Use snake_case for module names to avoid warning.
Also move utils modules to separate files.
